Tried to get _backtrack_coordinates() to work
In many cases its only colliding on one side so I thought I could get
the sides it WAS colliding on and move it along the other side, however
it appears that it actually is colliding on TWO sides, but not
necessarily sides that matter. i.e. if you collide on the side furthest
away from the center, then we don't care since we're moving away from
that side.

But this is all moot if I move to hierarchical bounding boxes